backend/app/requirements.txt
fastapi
uvicorn
sqlalchemy
databases
asyncpg
passlib[bcrypt]
pyjwt
python-dotenv
Flask
psycopg2-binary
python-jose
